feat/#57: component 구조 정리
