+++ app/(main)/experience/components/ExpList.tsx
@@ -24,7 +24,10 @@ export default function ExpList({ data }: ExpListProps) {
       <h1 className="text-[25px] font-bold mb-10 flex items-center justify-between">
         <span>내 경험</span>
         <div className="flex justify-end gap-4">
-          <BtnExp href="/addExperience " className="bg-primary-50 text-black">
+          <BtnExp
+            href="/addExperience "
+            className="bg-primary-50 text-gray-1000"
+          >
             경험 추가
           </BtnExp>
           <BtnFilter onSelectType={setSelectedType} />
